Hospital/requirements.txt
django
djangorestframework
gunicorn
coverage
django-heroku
django-cors-headers
mysqlclient
requests
numpy